refactor: refactor code structure
